finished generating teh leaderboard datasets
