+++ notebooks/Makefile
@@ -1,0 +1,2 @@
+leaderboard:
+	python3 makeLeaderboardDataset.py --inputFolder ../data --outputFolder ../data
