--- scripts/Makefile
@@ -1,2 +1,0 @@
-leaderboard:
-	python3 makeLeaderboardDataset.py --spreadsheetFolder ../../neil_repo
